add support for photo transformations in shotwell
Shotwell has support for various image transformations like cropping,
straightening, color adjustments, etc. Add support for using imagemagick
to perform some of these adjustments. This creates a new 'transformed'
directory in the photo HTML output directory so that the original image
is preserved.

Note that not all of the photo operations are currently supported.

Signed-off-by: Brian Masney <masneyb@onstation.org>