add optional support for converting videos to a common format
Add the optional --video-convert-command and --video-convert-ext
arguments that allow using an external command to convert videos into a
common format for consistency across the rendered library. The original
videos are preserved and the converted videos are stored in the
transformed directory.

While changes are being made here, let's also pass in the imagemagick
command as well so it can be overridden if needed.

Signed-off-by: Brian Masney <masneyb@onstation.org>