fix scaled videos that are rotated
Some 480p/720p/1080p-scaled videos that the original was rotated do not
have the correct rotation. Fix this so that the scaled videos also have
the correct rotation.

Note that this also changes it so that the original video is used since
that's where the correct rotation tag is (if present). This may cause
your video files to be regenerated unnecessarily. You can avoid this by
updating the corresponding *.idx files stored next to the videos if you
don't want to regenerate the media.

Signed-off-by: Brian Masney <masneyb@onstation.org>